+++ apps/manifest-demo/webpack-host/webpack.config.js
@@ -1,4 +1,6 @@
 const path = require('path');
+const reactPath = path.dirname(require.resolve('react/package.json'));
+const reactDomPath = path.dirname(require.resolve('react-dom/package.json'));
 // const { registerPluginTSTranspiler } = require('nx/src/utils/nx-plugin.js');
 // registerPluginTSTranspiler();
 const {
@@ -33,7 +35,6 @@ module.exports = composePlugins(withNx(), withReact(), (config, context) => {
           'rspack_manifest_provider@http://localhost:3011/mf-manifest.json',
         'js-entry-provider':
           'rspack_js_entry_provider@http://localhost:3012/remoteEntry.js',
-        'modern-js-provider': 'app1@http://127.0.0.1:4001/mf-manifest.json',
       },
       filename: 'remoteEntry.js',
       shared: {
@@ -65,6 +66,17 @@ module.exports = composePlugins(withNx(), withReact(), (config, context) => {
     }),
   );
 
+  config.plugins.push({
+    name: 'nx-dev-webpack-plugin',
+    apply(compiler) {
+      compiler.options.devtool = false;
+      compiler.options.resolve.alias = {
+        ...compiler.options.resolve.alias,
+        react: reactPath,
+        'react-dom': reactDomPath,
+      };
+    },
+  });
   config.plugins.forEach((p) => {
     if (p.constructor.name === 'ModuleFederationPlugin') {
       //Temporary workaround - https://github.com/nrwl/nx/issues/16983