+++ packages/rsbuild-plugin/src/utils/ssr.ts
@@ -3,7 +3,13 @@ import { createRequire } from 'node:module';
 import { encodeName } from '@module-federation/sdk';
 import { CALL_NAME_MAP } from '../constant';
 
-import type { EnvironmentConfig, RsbuildConfig, Rspack } from '@rsbuild/core';
+import type {
+  EnvironmentConfig,
+  RsbuildConfig,
+  Rspack,
+  EnvironmentContext,
+  DistPathConfig,
+} from '@rsbuild/core';
 import type { moduleFederationPlugin } from '@module-federation/sdk';
 
 const require = createRequire(import.meta.url);
@@ -11,6 +17,7 @@ const resolve = require.resolve;
 
 export const SSR_DIR = 'ssr';
 export const SSR_ENV_NAME = 'mf-ssr';
+export const ENV_NAME = 'mf';
 
 export function setSSREnv() {
   process.env['MF_SSR_PRJ'] = 'true';
@@ -19,6 +26,30 @@ export function setSSREnv() {
 const isDev = () => {
   return process.env['NODE_ENV'] === 'development';
 };
+
+export function patchNodeConfig(
+  config: Rspack.Configuration,
+  mfConfig: moduleFederationPlugin.ModuleFederationPluginOptions,
+) {
+  config.output ||= {};
+  config.target = 'async-node';
+  // @module-federation/node/universe-entry-chunk-tracker-plugin only export cjs
+  const UniverseEntryChunkTrackerPlugin =
+    require('@module-federation/node/universe-entry-chunk-tracker-plugin').default;
+  config.plugins ||= [];
+  isDev() && config.plugins.push(new UniverseEntryChunkTrackerPlugin());
+
+  const uniqueName = mfConfig.name || config.output?.uniqueName;
+  const chunkFileName = config.output.chunkFilename;
+  if (
+    typeof chunkFileName === 'string' &&
+    uniqueName &&
+    !chunkFileName.includes(uniqueName)
+  ) {
+    const suffix = `${encodeName(uniqueName)}-[contenthash].js`;
+    config.output.chunkFilename = chunkFileName.replace('.js', suffix);
+  }
+}
 
 export function patchSSRRspackConfig(
   config: Rspack.Configuration,
@@ -46,25 +77,27 @@ export function patchSSRRspackConfig(
     // set virtue entry, only need mf entry
     config.entry = 'data:application/node;base64,';
   }
-  config.target = 'async-node';
-  // @module-federation/node/universe-entry-chunk-tracker-plugin only export cjs
-  const UniverseEntryChunkTrackerPlugin =
-    require('@module-federation/node/universe-entry-chunk-tracker-plugin').default;
-  config.plugins ||= [];
-  isDev() && config.plugins.push(new UniverseEntryChunkTrackerPlugin());
-
-  const uniqueName = mfConfig.name || config.output?.uniqueName;
-  const chunkFileName = config.output.chunkFilename;
-  if (
-    typeof chunkFileName === 'string' &&
-    uniqueName &&
-    !chunkFileName.includes(uniqueName)
-  ) {
-    const suffix = `${encodeName(uniqueName)}-[contenthash].js`;
-    config.output.chunkFilename = chunkFileName.replace('.js', suffix);
-  }
-
+  patchNodeConfig(config, mfConfig);
   return config;
+}
+
+export function patchToolsTspack(
+  envConfig: EnvironmentConfig,
+  fn: (
+    config: Rspack.RspackOptions,
+    { environment }: { environment: EnvironmentContext },
+  ) => void,
+) {
+  const rspackArr = [];
+  if (envConfig.tools?.rspack) {
+    if (Array.isArray(envConfig.tools?.rspack)) {
+      rspackArr.push(...envConfig.tools?.rspack);
+    } else if (typeof envConfig.tools?.rspack === 'function') {
+      rspackArr.push(envConfig.tools?.rspack);
+    }
+  }
+  envConfig.tools ||= {};
+  envConfig.tools.rspack = [...rspackArr, fn];
 }
 
 export function createSSRREnvConfig(
@@ -74,27 +107,33 @@ export function createSSRREnvConfig(
   rsbuildConfig: RsbuildConfig,
   callerName?: string,
 ) {
+  const rspackArr = [];
+  if (envConfig.tools?.rspack) {
+    if (Array.isArray(envConfig.tools?.rspack)) {
+      rspackArr.push(...envConfig.tools?.rspack);
+    } else if (typeof envConfig.tools?.rspack === 'function') {
+      rspackArr.push(envConfig.tools?.rspack);
+    }
+  }
   const ssrEnvConfig: EnvironmentConfig = {
     ...envConfig,
-    tools: {
-      rspack: (config, { environment }) => {
-        if (environment.name !== SSR_ENV_NAME) {
-          return;
-        }
-        patchSSRRspackConfig(config, mfConfig, ssrDir, callerName);
-      },
-    },
   };
+  patchToolsTspack(ssrEnvConfig, (config, { environment }) => {
+    if (environment.name !== SSR_ENV_NAME) {
+      return;
+    }
+    patchSSRRspackConfig(config, mfConfig, ssrDir, callerName);
+  });
   ssrEnvConfig.output = {
     ...ssrEnvConfig.output,
     // https://rsbuild.rs/config/output/target#other-targets
     // Rsbuild not support all rspack targets, so modify to async-node in modifyRspackConfig
     target: 'node',
     distPath: {
-      ...ssrEnvConfig.output?.distPath,
+      ...(ssrEnvConfig.output?.distPath as DistPathConfig),
       root: path.join(
-        ssrEnvConfig.output?.distPath?.root ||
-          rsbuildConfig.output?.distPath?.root ||
+        (ssrEnvConfig.output?.distPath as DistPathConfig)?.root ||
+          (rsbuildConfig.output?.distPath as DistPathConfig)?.root ||
           '',
         ssrDir,
       ),
@@ -104,33 +143,42 @@ export function createSSRREnvConfig(
   return ssrEnvConfig;
 }
 
-export function createSSRMFConfig(
+export function patchNodeMFConfig(
   mfConfig: moduleFederationPlugin.ModuleFederationPluginOptions,
 ) {
-  const ssrMFConfig = {
-    ...mfConfig,
-    exposes: { ...mfConfig.exposes },
-    library: {
-      ...mfConfig.library,
-      name: mfConfig.name,
-      type: mfConfig.library?.type ?? 'commonjs-module',
-    },
-    dts: false,
-    dev: false,
-    runtimePlugins: [...(mfConfig.runtimePlugins || [])],
+  if (mfConfig.remotes) {
+    mfConfig.remoteType = 'script';
+  }
+  mfConfig.exposes = { ...mfConfig.exposes };
+  mfConfig.library = {
+    ...mfConfig.library,
+    name: mfConfig.name,
+    type: mfConfig.library?.type ?? 'commonjs-module',
   };
+  mfConfig.runtimePlugins = [...(mfConfig.runtimePlugins || [])];
 
-  ssrMFConfig.runtimePlugins.push(
+  mfConfig.runtimePlugins.push(
     resolve('@module-federation/node/runtimePlugin'),
   );
   if (isDev()) {
-    ssrMFConfig.runtimePlugins.push(
+    mfConfig.runtimePlugins.push(
       // @ts-ignore
       resolve(
         '@module-federation/node/record-dynamic-remote-entry-hash-plugin',
       ),
     );
   }
+}
+
+export function createSSRMFConfig(
+  mfConfig: moduleFederationPlugin.ModuleFederationPluginOptions,
+) {
+  const ssrMFConfig = {
+    ...mfConfig,
+  };
+  patchNodeMFConfig(ssrMFConfig);
+  ssrMFConfig.dts = false;
+  ssrMFConfig.dev = false;
 
   return ssrMFConfig;
 }