+++ packages/rspack/src/ModuleFederationPlugin.ts
@@ -373,3 +373,8 @@ export class ModuleFederationPlugin implements RspackPluginInstance {
 }
 
 export const GetPublicPathPlugin = RemoteEntryPlugin;
+
+export {
+  TreeShakingSharedPlugin,
+  type TreeShakingSharedPluginOptions,
+} from './TreeShakingSharedPlugin';