+++ packages/bridge/bridge-react-webpack-plugin/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @module-federation/bridge-react-webpack-plugin
 
+## 2.0.0
+
+### Patch Changes
+
+- @module-federation/sdk@2.0.0
+
+## 0.24.1
+
+### Patch Changes
+
+- @module-federation/sdk@0.24.1
+
+## 0.24.0
+
+### Patch Changes
+
+- @module-federation/sdk@0.24.0
+
+## 0.23.0
+
+### Patch Changes
+
+- @module-federation/sdk@0.23.0
+
+## 0.22.1
+
+### Patch Changes
+
+- @module-federation/sdk@0.22.1
+
+## 0.22.0
+
+### Patch Changes
+
+- @module-federation/sdk@0.22.0
+
 ## 0.21.6
 
 ### Patch Changes