+++ packages/chrome-devtools/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @module-federation/devtools
 
+## 2.0.0
+
+### Major Changes
+
+- 2bd445d: feat: upgrade to modern js 3.0
+
+### Patch Changes
+
+- @module-federation/sdk@2.0.0
+
+## 0.24.1
+
+### Patch Changes
+
+- @module-federation/sdk@0.24.1
+
+## 0.24.0
+
+### Patch Changes
+
+- @module-federation/sdk@0.24.0
+
+## 0.23.0
+
+### Patch Changes
+
+- 813923c: feat(chrome-devtools): support i18n and dark mode
+  - @module-federation/sdk@0.23.0
+
+## 0.22.1
+
+### Patch Changes
+
+- 88bee80: feat(chrome-devtools): optimize UI
+  - @module-federation/sdk@0.22.1
+
+## 0.22.0
+
+### Patch Changes
+
+- @module-federation/sdk@0.22.0
+
 ## 0.21.6
 
 ### Patch Changes