+++ packages/create-module-federation/CHANGELOG.md
@@ -1,5 +1,21 @@
 # create-module-federation
 
+## 2.0.0
+
+### Major Changes
+
+- 2bd445d: feat: upgrade to modern js 3.0
+
+## 0.24.1
+
+## 0.24.0
+
+## 0.23.0
+
+## 0.22.1
+
+## 0.22.0
+
 ## 0.21.6
 
 ## 0.21.5