+++ packages/esbuild/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @module-federation/esbuild
 
+## 0.0.95
+
+### Patch Changes
+
+- @module-federation/sdk@2.0.0
+
+## 0.0.94
+
+### Patch Changes
+
+- @module-federation/sdk@0.24.1
+
+## 0.0.93
+
+### Patch Changes
+
+- @module-federation/sdk@0.24.0
+
+## 0.0.92
+
+### Patch Changes
+
+- @module-federation/sdk@0.23.0
+
+## 0.0.91
+
+### Patch Changes
+
+- @module-federation/sdk@0.22.1
+
+## 0.0.90
+
+### Patch Changes
+
+- @module-federation/sdk@0.22.0
+
 ## 0.0.89
 
 ### Patch Changes