+++ packages/metro-plugin-rnc-cli/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @module-federation/metro-plugin-rnc-cli
 
+## 2.0.0
+
+### Patch Changes
+
+- @module-federation/metro@2.0.0
+
+## 0.24.1
+
+### Patch Changes
+
+- @module-federation/metro@0.24.1
+
+## 0.24.0
+
+### Patch Changes
+
+- @module-federation/metro@0.24.0
+
+## 0.23.0
+
+### Patch Changes
+
+- @module-federation/metro@0.23.0
+
+## 0.22.1
+
+### Patch Changes
+
+- @module-federation/metro@0.22.1
+
+## 0.22.0
+
+### Patch Changes
+
+- @module-federation/metro@0.22.0
+
 ## 0.21.6
 
 ### Patch Changes