+++ packages/nextjs-mf/CHANGELOG.md
@@ -1,5 +1,66 @@
 # @module-federation/nextjs-mf
 
+## 8.8.55
+
+### Patch Changes
+
+- @module-federation/enhanced@2.0.0
+- @module-federation/node@2.7.31
+- @module-federation/runtime@2.0.0
+- @module-federation/webpack-bundler-runtime@2.0.0
+- @module-federation/sdk@2.0.0
+
+## 8.8.54
+
+### Patch Changes
+
+- @module-federation/runtime@0.24.1
+- @module-federation/node@2.7.30
+- @module-federation/webpack-bundler-runtime@0.24.1
+- @module-federation/enhanced@0.24.1
+- @module-federation/sdk@0.24.1
+
+## 8.8.53
+
+### Patch Changes
+
+- Updated dependencies [c3c068e]
+  - @module-federation/enhanced@0.24.0
+  - @module-federation/node@2.7.29
+  - @module-federation/runtime@0.24.0
+  - @module-federation/webpack-bundler-runtime@0.24.0
+  - @module-federation/sdk@0.24.0
+
+## 8.8.52
+
+### Patch Changes
+
+- @module-federation/enhanced@0.23.0
+- @module-federation/runtime@0.23.0
+- @module-federation/node@2.7.28
+- @module-federation/webpack-bundler-runtime@0.23.0
+- @module-federation/sdk@0.23.0
+
+## 8.8.51
+
+### Patch Changes
+
+- @module-federation/runtime@0.22.1
+- @module-federation/enhanced@0.22.1
+- @module-federation/webpack-bundler-runtime@0.22.1
+- @module-federation/sdk@0.22.1
+- @module-federation/node@2.7.27
+
+## 8.8.50
+
+### Patch Changes
+
+- @module-federation/runtime@0.22.0
+- @module-federation/enhanced@0.22.0
+- @module-federation/webpack-bundler-runtime@0.22.0
+- @module-federation/sdk@0.22.0
+- @module-federation/node@2.7.26
+
 ## 8.8.49
 
 ### Patch Changes