+++ packages/runtime-plugins/inject-external-runtime-core-plugin/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @module-federation/inject-external-runtime-core-plugin
 
+## 2.0.0
+
+### Patch Changes
+
+- @module-federation/runtime-tools@2.0.0
+
+## 0.24.1
+
+### Patch Changes
+
+- @module-federation/runtime-tools@0.24.1
+
+## 0.24.0
+
+### Patch Changes
+
+- @module-federation/runtime-tools@0.24.0
+
+## 0.23.0
+
+### Patch Changes
+
+- @module-federation/runtime-tools@0.23.0
+
+## 0.22.1
+
+### Patch Changes
+
+- @module-federation/runtime-tools@0.22.1
+
+## 0.22.0
+
+### Patch Changes
+
+- @module-federation/runtime-tools@0.22.0
+
 ## 0.21.6
 
 ### Patch Changes