+++ packages/runtime/CHANGELOG.md
@@ -1,5 +1,55 @@
 # @module-federation/runtime
 
+## 2.0.0
+
+### Patch Changes
+
+- @module-federation/sdk@2.0.0
+- @module-federation/error-codes@2.0.0
+- @module-federation/runtime-core@2.0.0
+
+## 0.24.1
+
+### Patch Changes
+
+- Updated dependencies [6590738]
+  - @module-federation/runtime-core@0.24.1
+  - @module-federation/sdk@0.24.1
+  - @module-federation/error-codes@0.24.1
+
+## 0.24.0
+
+### Patch Changes
+
+- @module-federation/sdk@0.24.0
+- @module-federation/error-codes@0.24.0
+- @module-federation/runtime-core@0.24.0
+
+## 0.23.0
+
+### Patch Changes
+
+- Updated dependencies [3aa3cf9]
+  - @module-federation/runtime-core@0.23.0
+  - @module-federation/sdk@0.23.0
+  - @module-federation/error-codes@0.23.0
+
+## 0.22.1
+
+### Patch Changes
+
+- @module-federation/sdk@0.22.1
+- @module-federation/error-codes@0.22.1
+- @module-federation/runtime-core@0.22.1
+
+## 0.22.0
+
+### Patch Changes
+
+- @module-federation/sdk@0.22.0
+- @module-federation/error-codes@0.22.0
+- @module-federation/runtime-core@0.22.0
+
 ## 0.21.6
 
 ### Patch Changes