+++ packages/storybook-addon/CHANGELOG.md
@@ -1,5 +1,54 @@
 # @module-federation/storybook-addon
 
+## 6.0.0
+
+### Major Changes
+
+- 2bd445d: feat: upgrade to modern js 3.0
+
+### Patch Changes
+
+- @module-federation/enhanced@2.0.0
+- @module-federation/sdk@2.0.0
+
+## 5.0.6
+
+### Patch Changes
+
+- @module-federation/enhanced@0.24.1
+- @module-federation/sdk@0.24.1
+
+## 5.0.5
+
+### Patch Changes
+
+- f46cdd6: prefer use `stroybook/internal/node-logger` as logger
+- Updated dependencies [c3c068e]
+  - @module-federation/enhanced@0.24.0
+  - @module-federation/sdk@0.24.0
+
+## 5.0.4
+
+### Patch Changes
+
+- @module-federation/enhanced@0.23.0
+- @module-federation/sdk@0.23.0
+
+## 5.0.3
+
+### Patch Changes
+
+- @module-federation/enhanced@0.22.1
+- @module-federation/sdk@0.22.1
+
+## 5.0.2
+
+### Patch Changes
+
+- a1c1c51: prefer use `stroybook/internal/node-logger` as logger
+  - @module-federation/enhanced@0.22.0
+  - @module-federation/sdk@0.22.0
+
 ## 5.0.1
 
 ### Patch Changes